Fix periodic thread that monitors Ceph
A timeout between ceph-manager and sysinv is causing ceph-manager to
stop responding. When Ceph Manager detects that 'require_jewel_osds'
flag needs to be set it queries sysinv which for whatever reason
(most likely dealing with some final operations) fails to respond in
a reasonable amount of time. This causes an exception in ceph-manager
which breaks one of the periodic threads execution.

Change-Id: If49f5ffbce4aeac3d50d52f526d1ce905be3cecb
Signed-off-by: Kristine Bujold <kristine.bujold@windriver.com>
Signed-off-by: Scott Little <scott.little@windriver.com>