Fix bug in logmgmt start check
The start function of the logmgmt init script
checks for a valid pidfile to see if the process is
already running. Unfortunately, the code has a couple
of typos that cause the check to fail if the "start"
is called when the process is already running.

This commit fixes the typos.

Change-Id: I5795d23cc9e41a18b62e35bf3df07817522efe52
Related-Bug: 1827326
Signed-off-by: Don Penney <don.penney@windriver.com>