Add update-iso.sh for user customization of ISO
This commit adds a new tool, update-iso.sh, that allows
a user to customize an ISO to add or modify installation
boot parameters or to add a custom kickstart post-install
addon.

For example, to modify the boot_device and rootfs_device
flags to use a different disk from the default, or to use
the persistent by-path symlink, the user can run:

update-iso.sh -i export/bootimage.iso \
    -o updated/bootimage.iso \
    -p rootfs_device=/dev/disk/by-path/pci-0000:00:0d.0-ata-1.0 \
    -p boot_device=/dev/disk/by-path/pci-0000:00:0d.0-ata-1.0

update-iso.sh -i export/bootimage.iso \
    -o updated/bootimage.iso \
    -p rootfs_device=nvme0n1 \
    -p boot_device=nvme0n1

Or to add a custom addon:
update-iso.sh -i export/bootimage.iso \
    -o updated/bootimage.iso \
    -a updated/ks-addon.cfg

Additionally, this utility allows the user to specify the default boot
menu item selection, along with a menu timeout, to customize an ISO
for a specific installation type. This allows the user to create a
bootable ISO that automatically initiates installation with desried
parameters.

The update-iso.sh utility is packaged in platform-utils-controller,
making it available on controller hosts. This will be used in a
distributed cloud configuration in order to generate a modified ISO
that can be used for installing a subcloud with specific initial
configuration.

Change-Id: Ie6c9d034179bc879b7af0969ff313742fffa9895
Story: 2006595
Task: 37260
Signed-off-by: Don Penney <don.penney@windriver.com>