Fix pylint zuul error due to python-libvirt version
Nov 6th saw the release of a new version of python-libvirt.
tox.ini did not have an upper-constraint set for it,
so we picked up that version, which caused zuul to fail
in pylint.

Adding an upper-constraint to fix the issue.

This allows all reviews in stx/utilities to pass zuul
and be able to merge.

Change-Id: Idcdf37fbfb2f8745aadff9aacba828e8166dbc46
Closes-Bug: 1851707
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>