Protect against parallel exec of gen-bootloader-iso.sh
The bulk of the work done by gen-bootloader-iso.sh is not safe to be
executed by two callers in parallel. If the initial execution is
setting up shared files when a subsequent call starts, the subsequent
call can end up with an incomplete set of symlinks to shared files.
This, in turn, can result in installation failures.

To protect against this, a file lock is added to gen-bootloader-iso.sh
to prevent subsequent calls from accessing shared files before they're
completely setup.

Change-Id: Id6def4527b226b8746b2b8ff74059c0d3937e130
Closes-Bug: 1885779
Signed-off-by: Don Penney <don.penney@windriver.com>