Use host patches for subcloud install setup
This update enhances gen-bootloader-iso.sh to use patches applied on
the host controller when setting up the subcloud installation
bootloader. This checks the release version from the ISO to look for
patches for that release only, copying applied and committed patches
to the installation setup.

This update also adds cleanup of partial setup on failure.

Change-Id: I4eeb0141b75e48ce53c132b89e19e07c68cc49d2
Story: 2007994
Task: 40631
Signed-off-by: Don Penney <don.penney@windriver.com>