Improve DNF cache handling in gen-bootloader-iso.sh
The "dnf repoquery" command is using and updating the cache for ad-hoc
repos created when using the repofrompath option. This can result in
gen-bootloader-iso.sh using stale cached information when looking for
specific RPMs in the patch repo when generating bootloader ISOs for
subcloud installs.

In order to avoid such issues, this update asks DNF to mark the cache
as expired after copying the patch repo for generating the new ISOs,
as well as on deletion. This ensures the repoquery uses up-to-date
information.

Additionally, this update removes a duplicated patch repo setup step,
which was not cleaned up during dev testing on the previous commit.

Change-Id: I664edb346692d79cd13b23772686ea88f0aaaf9f
Story: 2007994
Task: 40798
Signed-off-by: Don Penney <don.penney@windriver.com>