Fix gen-bootloader-iso.sh error on unpatched controller
On a system that has never been patched, the updates repository may
not yet have a Packages directory. This update fixes the patch setup
in gen-bootloader-iso.sh to check for this case and handle it without
failing.

This update also improves the platform-kickstarts-pxeboot check to
look for a pxeboot kickstart, to ensure there isn't a false-positive
match against an extracted pxe-network-installer patch, which would
also have a /pxeboot directory.

Change-Id: I8e24a3ff123f4c2649be52b1a0ce77d830e342f3
Closes-Bug: 1893990
Signed-off-by: Don Penney <don.penney@windriver.com>