Fix missing log_error in update-iso.sh
The update-iso.sh utility references an undefined log_error function.
As a result, error messages are not properly reported, instead
resulting in messages such as:
stx-iso-utils.sh: line 53: log_error: command not found

This error was introduced by code restructuring in a previous update:
https://review.opendev.org/707673

This update defines a log_error function for update-iso.sh that writes
the message to stderr, as this is a user-run utility, so no need to
log to syslog.

Change-Id: Ifdc2a25c24cd86e61ccc12601a0bcbe6888ac349
Closes-Bug: 1895042
Signed-off-by: Don Penney <don.penney@windriver.com>