Remove empty package from stx-extensions
Packages defined in a spec with no files do not result in an RPM
produced by the build. On a rebuild, the build tools scan the spec and
sees the package defined but does not find a corresponding RPM, and so
flags the package for a rebuild as a result.

This commit removes the empty package definition from the spec.

Change-Id: I48e725bee837cd27352d38732d6f7155827cfe32
Partial-Bug: 1910439
Signed-off-by: Don Penney <don.penney@windriver.com>