Fix clear passwords presented in collected log files
The --relative option used by rsync resulted in changes
to the paths of some collected files in the tarball,
which in turn resulted in failures to mask the passwords
in those files (as they were no longer in the expected
location).

This update removes the --relative option, restoring
the location to that expected by the masking.

Closes-Bug: 1906524

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
Change-Id: I001d3e2b1d2ec0ff88a129fe7b31bee21928686b