Update ceph-manager to recover TLS certificate
Code changes in client.py to recreate the TLS CA certificate
temporary file if it had been unexpectedly removed or not found.
Without this temporary file, ceph-manager can not check health of
ceph by ceph-api and raise Ceph-Down alarm even if the cluster were
operational.

The exception thrown by the request lib is an IOError raised in the
cert_verify method. The _refresh_session is a method that set the
certification file if it has not been set before. But if the
certification file has been set and after it were removed, the method
do not try to reset it. To avoid duplicated code, a flag has been
added to this method to force the recover of the TLS certification.

Closes-Bug: #1940471

Signed-off-by: Daniel Pinto Barros <DanielPinto.Barros@windriver.com>
Change-Id: Icb15337e1e76ae92cd62631190a6739960f6245d