Add a script to update registry credentials
We have a script in https://docs.starlingx.io/dist_cloud/kubernetes/\
updating-docker-registry-credentials-on-a-subcloud.html
to update the docker registry credentials on a subcloud. As this
script is expected to use in multiple scenarios, this commit adds this
script in the /usr/local/bin directory, so it can be called to update
the registry credentials.

Changes against the original script:
1. Add ghcr-registry as it is newly introduced.
2. Add "source /etc/platform/openrc", so the OpenStack environmental
variables can be included.
3. Prompt for input username and password if not provided.

Test:
1. Create a patch with the "platform-util-controller" and apply it on
an AIOSX controller.
2. Call the script with the sysinv username and password, check the
OpenStack secrets payload that the username and password are updated.
And the secrets' UUIDs are updated to service parameters.
3. Call the script without username and password, prompt for username
and password.
4. Call the script with 3 arguments, exit with an error message and
the usage.

Partial-Bug: 1947014
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>
Change-Id: I4d930b06992a22addb15f4d4edcfac31af5d440b