Created script to change private registry config
This script changes the needed system parameters in order to use a
different registry in an already deployed StarlingX. To do this, it
manipulates the system parameters 'url', 'type' and 'auth-secret' of
service 'docker'. Each one of these three parameters may exist inside
the docker service sections 'docker-registry', 'elastic-registry',
'gcr-registry', 'ghcr-registry', 'k8s-registry' and 'quay-registry'.

The tests below were executed using a StarlingX dev build.
The regression tests were not executed, as this script is not executed
unless the user calls it.

Test Plan:

PASS: In a AIO-SX deploy with no parameters present, execute the script
with 4 arguments (non-empty user and password), verify that the new
registry is used to download images.
PASS: In a AIO-SX deploy with all parameters present, execute the script
with 4 arguments (non-empty user and password), verify that the new
registry is used to download images.
PASS: In a AIO-SX deploy with no parameters present, execute the script
with 2 parameters (no user or password given), verify that the new
registry is used to download images.
PASS: In a AIO-SX deploy with all parameters present, execute the script
with 2 parameters (no user or password given), verify that the new
registry is used to download images.
PASS: In a distributed cloud scenario, reconfigure the central cloud
with a different private registry, verify that the new registry is used
to download images, then check that new images downloaded by central
cloud are being successfully downloaded in the subclouds.

Closes-Bug: 1948491
Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: I3e10486333964264c8754c519163083a821eb54e