Add body argument to enable_msgr2 call
The _request method is always deleting 'body' parameter from kwargs.
In case a given parameter do not have a body, it will raise an
exception.
The fix adds a 'body' parameter to enable_msgr2 so it can be deleted by
_request method.

This problem was found during upgrade of AIO-SX after method
enable_msgr2 was created. The test was made only for AIO-SX but this
solves the problem in all configs.

Testing performed:
 - Upgrade of AIO-SX

Story: 2009074
Task: 43989

Signed-off-by: Vinicius Lopes da Silva <vinicius.lopesdasilva@windriver.com>
Change-Id: I4d3949a29ad9800508a6833701eff927f8c1e76d