Exclude private certificate and key files from collect
Collecting private keys and certificates is objectionable.

This update adds /etc/kubernetes/pki to the collect
exclude list so that its private keys and certificate
files are not collected.

This update then enhances collect's mask passwords
plugin to pull in the public certificate files which
do not pose a security risk and are useful during debug.

Test Plan:

PASS: Verify only the .crt files from the newly excluded
      dir are collected.

Change-Id: Ia9e3e132d10fd31486b92742fb17147b36aa98c9
Closes-Bug: 1952676
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>