Skip the secret deletion if it doesn't exist
If the registry secret exists in the service parameters but doesn't
exist in the OpenStack secret list, the current update script will
throw out an error. This commit skips the secret deletion operation if
the secret doesn't exists, allows creating a new secret and updating
it to the service parameters.

Test:
1. Bring up a DC(DX central cloud + SX subcloud) with this change.
2. Delete the ghcr registry secret manually, update the secrets again
with sysinv credentials in the subcloud.
3. Delete the ghcr registry secret manually, migrate the subcloud to
another central cloud.

Closes-bug: 1953365
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>
Change-Id: Id6afea988fa8559d8a80eaa5ddeec0a35a386014