Fix the list of packages to be copied to prestage shared directory
The script is copying all the files over to the
prestage/shared/Packages directory, instead of only a subset of
files.

This fix makes the relevant change so that only files whose
md5sums are changed, will be copied over.

Tests:
PASS: Verify that only packages to be downloaded are copied to the
      prestage/shared/Packages directory

Closes-bug: 1961861
Change-Id: Iea0e219895c33b17c1ff4279d6d73fc65289006f
Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>