Removing remaining pci affinity agent service code
This code is being removed as part of an effort to move the
pci-irq-affinity-agent into a openstack application container, instead
of a platform service.

There is still some code in this repo that will be deprecated now that
the service will not run on the platform anymore and therefore needs to
be removed. This change should clean the repo from unused code related
to the pci-irq-affinity-agent and also ensure that the platform will
no longer have this service configured and activated.

TEST PLAN:
PASS: Build the pci-irq-affinity-agent rpm package
PASS: Build the pci-irq-affinity-agent wheels
PASS: Build the new container image for the agent
PASS: Build and Install a platform ISO
PASS: Check that the agent service is not present on the platform
PASS: Check that the python package was not installed on the platform
PASS: Check that the rpm is not present on the platform
PASS: Launch a VM and check the events being received on the
pci-irq-affinity-agent container

Story: 2009299
Task: 44607

Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: I3a1ed8a2a0b4aa8b14b42618da71e5124688f424