List of packages in current release not found during prestage package preparation
When the script prestage-prepare-packages.sh is run, it attempts
to locate a file containing the names of packages and their
checksums in a given release. The script would fail even if
the file exists in /usr/local/share/pkg-list. The solution is
to look for the exact file and not just for files ending in
_packages_list.txt.

Test Plan:
PASS: Verify that the file is correctly resolved when the script
      is run and the file exists at /usr/local/share/pkg-list.

Closes-Bug: 1962794
Change-Id: I273b5d70e3b3877ee3a016c88dfd09f7b94be2eb
Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>