debian: Update update-motd to 3.9
The way the the /etc/motd message is displayed to the user differs
between Centos and Debian. In Debian (versions 8+), the actual motd
is printed and displayed by the pam_motd pam module. In Centos,
the pam_motd module is installed but its not used.

In order to get backwards compatibility with Centos we are doing
a copy in drop version of the Centos 7 version of update-motd.
However, this does not work correctly due to the pam_motd use.
In order for this to work correctly on Debian, we dropped in the
Ubuntu version of update-motd debian packaging since there isnt
a Debian version that we can use.

Test Plan
PASS Build package and ISO
PASS Boot ISO
PASS Check for /etc/update-motd.d directory

Story: 2009101
Task: 43892

Signed-off-by: Chuck Short <charles.short@windriver.com>
Change-Id: If68a385d1886892a2da36a271be05ce2bd0d69ac