Adding temporary debian integration scripts
In order to bootstrap and unlock on debian, there are
several workarounds required while the code is being modified
to properly support both Centos and Debian.

This is considered a temporary folder because eventually all
the workarounds should end up being resolved and removed
from the scripts.

Updated tox to exclude the two scripts from bashate

Story: 2009101
Task: 44827
Co-Authored-By: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I0cca8e564c1f30333ef7d060ea7f8a6089187d1a