debian-integration: Update workaround for lvs
This work is part of Debian integration effort. This will enable
part of sanity tests to pass.

For a lock-unlock scenario it is observed that the ceph monitor will
use rootfs instead of LV for ceph monitor data. This causes data loss,
preventing OSDs from finishing the boot phase(going up).

During lock-unlock it is observed that the 'mount' resource provided
by puppet does not try to mount the entrypoint for ceph monitor LV.
In contrast it does try for others, and this(for others) could be
observed as early as initial unlock. This is why a workaround for
unmounting was introduced in the first place and is currently enabled.

Added code to make sure the entrypoints are mounted.
These 2 workarounds are a very good candidate for a proper fix.

Tests on AIO-SX.
On a VM which have OSDs failed after a lock-unlock test, restored the
snapshot before doing initial unlock. Applied the extra workaround
added in this commit.
PASS: unlock
PASS: lock & unlock, OSDs go up.

Story: 2009101
Task: 44827
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: I8d1a56c1047f5e7bd598d987f1e54e7dab110f2b