Enable overwrite existing installation with prestage iso content
When an existing installation exists on the subcloud, we may not
want to overwrite the installation. The current behaviour is to
avoid overwrites.

However, if the user so desires, the existing installation can
be overwritten by the usage of an argument to
gen-prestaged-iso.sh.

A bug is also fixed in this code checkin, where copy_to_iso
gets one less argument than it should, causing the script to
fail when generating a prestaged iso.

Test Plan:
PASS: Verify that the force_install option is present on the
      kernel command line in the generated prestaged iso when
      the script is invoked with the --force-install argument.

PASS: Verify that the force_install option is not present on
      the kernel command line in the default case

PASS: Verify that the prestaged iso installs the prestaged
      content when generated with the --force-install argument.

Story: 2009948
Task: 44902
Depends-On: https://review.opendev.org/c/starlingx/metal/+/836215

Change-Id: I9da3b8315dc04bfa8a95f4b0a020a9c7641c0e56
Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>