Debian: Add a workaround to populate INSTALL_UUID
The INSTALL_UUID is generated by kickstarts and stored
in the feeds directory, and also referenced in the
platform.conf file.

The kickstart script is not currently being invoked on
Debian, so this workaround manually does those steps
so that patching UUID checks will pass, and patch alarms
will not be raised after provisioning a debian system.

Test Plan:
  Verify Build, Bootstrap and Unlock Debian ISO
  Verify after lock/unlock no patch alarm is raised
  Verify in patching.log that the uuid check passes
 after unlock.
  
Story: 2009969
Task: 44983
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I6310ba0fcfd5ef5fda0a07def2ea51f6f7217e35