Customize prestaging through an external set-up script in the iso for use by kickstarts
To customize prestaging, a script can be provided to the
kickstarts via the prestage iso. This script is provided
to gen-prestage-iso.sh via a setup option.

When the prestage content installation starts, the kickstarts
use the embedded script to customize the rootfs and bootfs
devices.

Test Plan:

PASS: Verify the prestage iso functionality when created with
      the setup option and without the force-install option.

PASS: Verify the prestage iso functionality when created with
      the setup option and with the force-install option.

Regression:

PASS: Verify the prestage iso functionality when created
      without the setup and force-install options.

Depends-On: https://review.opendev.org/c/starlingx/metal/+/837758

Story: 2009948
Task: 45049

Change-Id: I7ae2716af8ab50773d1839a5e2509f0bf464298d
Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>