Debian: Partially re-add workaround for fminit
The fm_db_sync_event_suppression.py must be located
under the /usr/local/bin folder.

It imports fm_log which must be co-located.

It is referenced explicitly by its full PATH in one
of the FM constants files, and also in the helm charts.

Changes to the debian build are required to allow this
file to be installed to that location explicitly.

Test Plan:
  Added these lines to the workaround on a broken Debian
lab which allowed it to complete bootstrap.

Story: 2009101
Task: 44827
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: Iab1145f1ac8094e3c03f7ce55a7b1fcd9baee7d9