Revert "Debian: remove grubby workarounds"
This reverts commit 88a5cb29e1f7f44b8bb76d065213d2142493dab7.

Apparently the grub.cfg file cannot be modified at runtime,
else the stale signature will cause UEFI boot to fail.
This means that as it stands, grubby can't be allowed to run
and change the grub.cfg file.  So we put the workarounds back
in place until a proper solution is devised.

Story: 2009964
Task: 45126

Change-Id: Idb4d6e5645c3f5f6983c4d3b66a6d5d24e4fb677
Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>