Removing remaining pci affinity agent service code
This code is being removed as part of an effort to move the
pci-irq-affinity-agent into a openstack application container, instead
of a platform service.

There was a previous change [1] to clean the CentOS builds from the pci
affinity agent but it was reverted since it broke the Debian build.
Now that the Debian related changes [2] are already merged we can bring
this change back.

There is still some code in this repo that will be deprecated now that
the service will not run on the platform anymore and therefore needs to
be removed. This change should clean the repo from unused code related
to the pci-irq-affinity-agent and also ensure that the platform will
no longer have this service configured and activated.

TEST PLAN:
PASS: Build the pci-irq-affinity-agent rpm package
PASS: Build the pci-irq-affinity-agent wheels
PASS: Build the new container image for the agent
PASS: Build and Install a platform ISO
PASS: Check that the agent service is not present on the platform
PASS: Check that the python package was not installed on the platform
PASS: Check that the rpm is not present on the platform

Story: 2009299
Task: 44607

[1] https://review.opendev.org/c/starlingx/utilities/+/830892
[2] https://review.opendev.org/c/starlingx/utilities/+/840229

Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: I8b2286bff66f8e6f85dc0408459103b3b139528d