Re-apply removal of UAR 9
Remove unlock workaround #9 which disables puppet
configuration of ntp.

This has been done previously by [1], but was reverted
due to the change failing unlock tests in one of the ipv6
labs.

Failure reason was narrowed down to a new behavior
in which failing to sync the system clock to an ntp server
would cause unlock to fail.

An adjustment [2] has been made to correct this.

Test Plan:
PASS bootstrapped and unlocked
PASS ntp service running and enabled
PASS no puppet errors related to ntp
PASS ntp-modify updates ntp.conf
PASS clock syncs to ntp server when able to connect to it
PASS unlock does not fail when not getting a response
     from one of the net servers

Story: 2009965
Task: 45444

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/841784

References:
[1] https://review.opendev.org/c/starlingx/utilities/+/839419
[2] https://review.opendev.org/c/starlingx/stx-puppet/+/841784

Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>
Change-Id: I8fa516941f88dcc067e7a0a4c7154b7aa136be4e