Updating gen-bootloader-iso.sh to handle ostree
We need ostree to install a subcloud based on the Debian OS.

In this feature, support is provided to generate a debian miniboot
iso for subcloud installation over redfish in
gen-bootloader-iso.sh. The end result is a bootimage.iso that
will pull the ostree repo from the system controller when
it is inserted into the virtual media slot on the BMC of the
subcloud over Redfish.

Test Plan:

PASS: Verify that the bootimage.iso is able to install Debian
      successfully on the subcloud Virtual Machine (in
      VirtualBox).

PASS: Verify that the bootimage.iso is able to install CentOS
      successfully on a hardware subcloud

Story: 2010118
Task: 45738

Depends-On: https://review.opendev.org/c/starlingx/tools/+/848380
Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>
Change-Id: Ie2708266419db3607dc1c8fdbb70e638d75a0813