OPAE tools in container image stx-debian-tools-dev
The Open Programmable Acceleration Engine (OPAE) tools, and its
required dependencies, are introduced in the stx-debian-tools-dev
container, for the vRAN enablement with Intel hardware.

The OPAE tools deliverables consist of the following packages, which are
derived from the same source, but handled as individual packages. The
OPAE Intel FGPA driver is not required in the container, because it is
already supported in StarlingX kernel.

  - opae-sdk-1.3.7-5
  - python3-opae.admin-1.0.3
  - python3-opae.pacsign-1.0.4
  - opae-intel-fpga-driver-2.0.1

When building OPAE from upstream source code, binaries are installed in
bin directory, though privileged commands should be located in sbin.
According to the upstream documentation, most OPAE commands need to be
run in privileged mode, but the description is not always clear.
Hence, in order to provide a consistent CLI interface in StarlingX, the
OPAE binaries are installed in the proper standard directories.

Test Plan:
  - Container image is built successfully on Debian (PASS)
  - OPAE packages are successfully installed (PASS)
  - Run time testing of the OPAE tools (WIP)

Story: 2010138
Task: 46170

Depends-On: https://review.opendev.org/c/starlingx/tools/+/862743

Signed-off-by: Rogerio Ferraz <RogerioOliveira.Ferraz@windriver.com>
Change-Id: I9430fb34362ee487b0de76b4dbef78331446f5f3