Adding prestaging iso creation script for Debian
Factory pre-stage and infield installation are supported via an
iso which is used for local installation. This iso installs the ostree
repo on /opt/platform-backup and a basic Debian installation on the
machine. This iso is generated with a script: gen-prestage-iso.sh.

If there is already a subcloud installation on the machine, then the
script will not install the ostree repo or Debian.

However, with the force-install option, the iso will install the repo
and Debian on the machine.

Test Plan:

PASS: Verify that the prestage iso is able to install the ostree repo
      and Debian on a subcloud with no existing installation

PASS: Verify that the prestage iso fails to install when there is an
      existing installation on the subcloud

PASS: Verify that the prestage iso installs on the subcloud even when
      there is an existing installation when the force-install option
      is provided to the iso creation script (gen-prestage-iso.sh)

PASS: Verify that the prestage iso installs on the subcloud with
      sata and nvme storage media

PASS: Verify that a single prestage iso can be used to install
      irrespective of the media type, when configured with ks-setup.cfg.

Depends-On: https://review.opendev.org/c/starlingx/metal/+/853545
Story: 2010120
Task: 46134

Change-Id: Ic5cdaa694b4d4223610f419aed456e3c8e851e9a
Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>