debian_iso_image.inc: correct the install package name of utilities
Currently we use the "stx-std.lst" file in tools to add install package.
We are going to use the "debian_iso_image.inc" file to follow CentOS.

We observed some layers already have the debian_iso_image.inc, so first
of all, we need to check if all the packages are available, otherwise,
once supporting "debian_iso_image.inc", "build-image" will fail.

The binary name of python-cephclient is changed to python3-cephclient
built with python3. The python3-cephclient can be found in
"stx-std.lst".

Story: 2008846
Task: 46141

Signed-off-by: Yue Tao <yue.tao@windriver.com>
Change-Id: I15b57a5561fd2524690fdc353353d2c369b387d9