K8s Coredump Handler: initial package
Currently, it is not possible to set the core_pattern [1] on a per Pod
basis. That is, the kernel.core_pattern sysctl is not namespaced and
applies to the entire node.

Applications demands exclusive access to the core dump data for
debugging and support without enabling full access to the host OS files.

In order to meet both the platform and application requirements, a new
Kubernetes aware core dump handler is proposed.

This review covers the initial code and packaging of this handler, which
still represents limited functionality, only applying to very basic
testing.

Reference to coredump configuration:
1. https://man7.org/linux/man-pages/man5/coredump.conf.5.html

Test Plan:
PASS: Basic functionality after bootstrap (call handler directly)
PASS: Check handler is called upon coredump (core_pattern config)

Story: 2010261
Task: 46156
Depends-On: https://review.opendev.org/c/starlingx/integ/+/854682
Co-Authored-By: Samuel Presa Toledo <samuel.presatoledo@windriver.com>
Co-Authored-By: Heron Vieira <heron.vieira@windriver.com>

Signed-off-by: Adriano Oliveira <adriano.oliveira@windriver.com>
Change-Id: I7e18b27a6de1e74d8df3e4c28661959763ce7468