K8s-coredump handler - Annotations configuration
Parse, validate and apply configurations for the following annotations:
starlingx.io/core_compression: If configured, should accept None or lz4
starlingx.io/core_max_size: Parse, validate and apply accordingly to
    coredump configuration ProcessSizeMax
starlingx.io/core_max_used: Parse, validate and apply accordingly to
    coredump configuration MaxUse
starlingx.io/core_min_free: Parse, validate and apply accordingly to
    coredump configuration KeepFree
Reference to coredump configuration:
1. https://man7.org/linux/man-pages/man5/coredump.conf.5.html

It's important to notice that the configuration file
(/etc/k8s-coredump-conf.json) is going to be implemented
on the next task for this epic, so to run the handler
as is in this review, it's necessary to manually configure
this file with the kubectl token.
Is also important to notice that the k8s-coredump-handler
is not configured by default, since this configuration
is set to be done in another task.

Test Plan:
PASS: Verify if configuring the annotation
  starlingx.io/core_compression as lz4 will compress the coredump file
PASS: Verify if the system validates the max file size with the
  annotation starlingx.io/core_max_size
PASS: Verify if the system validates the space used with the
  annotation starlingx.io/core_max_used
PASS: Verify if the system validates the free space with the
  annotation starlingx.io/core_min_free
PASS: Verify if the logs are being generated

Story: 2010261
Task: 46172

Signed-off-by: Heron Vieira <heron.vieira@windriver.com>
Change-Id: I909ec625f1e0fb031ec28ee6b5df9f1dc65670ce