Use hostname on restful service certificate
Force the use of hostname on getting the certificates
for restful service, instead of the IP address.
This is necessary due to the update of the Ceph version
in rook-ceph app. New Ceph version started to store the IP
number instead of the hostname in restful service infos.
We need a new function to return the hostname in order to
get the key and certificate files.

Story: 2009138
Task: 46735

Test Plan:
PASS: AIO-SX - created an image with this version of code
and use it to get service certificate

Signed-off-by: Daian Cardoso Sganderlla <Daian.CardosoSganderlla@windriver.com>
Change-Id: I06ac2b40662ef1df4b86d82d2cdf298d746d886b