Wrong debian controller root disk I/O scheduler
On All-in-One systems and controller systems, Debian-based STX
does not set the I/O scheduler of the disk hosting the root file
system to the BFQ (Budget-Fair Queuing) I/O scheduler.

The I/O scheduler was not set on Debian because is-rootdisk-device.sh
script was not detecting the root disk correctly due to ostree.

Code change example:
Existing CentOS code running on Debian
[root@controller-0 ~(keystone_admin)]# findmnt -n -T / -o SOURCE
/dev/sda5[/ostree/deploy/debian/deploy/bcd4ae00dd670c967b62c7e6e471f9899174f81ab7e379105901e5e88f2aba44.1]

After modification:
[root@controller-0 ~(keystone_admin)]# findmnt -n -T / -o SOURCE | cut -d'[' -f1
/dev/sda5

Test Plan:
PASS: It was checked disk I/O scheduler is now 'bfq' in
AIO-SX fresh install

Closes-Bug: 1995739
Change-Id: Ia08335ce73c565aebb97ea85e32cfd543c4594bc
Signed-off-by: Pedro Vinícius Silva da Cruz
<pedro.silvadacruz@windriver.com>