Specify a replacement for the installer kickstart
This commit allows for an optional replacement for
the installer kickstart. This kickstart will not replace
the actual kickstart in the runtime - it is only used
at the time of the prestage iso install.

Also included is an additional check for the boot_device
and the rootfs_device. It is possible that only one of them
is specified, and this may result in instdev not being set,
leading to a failure in the installation of the prestaged iso.

Test Plan:
PASS: Verify the the replacement kickstart is used for
      installation when specified via the command line.

PASS: Verify that passing in either rootfs_device or
      boot_device in the params argument (but not both) will
      be sufficient to install the prestage iso.

Story: 2009948
Task: 46904

Signed-off-by: ssharma3 <shrikumar.sharma@windriver.com>
Change-Id: I84becdce9f094e481fb911fb463951c4ffbd8130