Adding retry when getting ceph services in cephmgr
After bootstrap system, it takes a while to provide ceph
services available. This can cause problems when updating
via nfv for example.
To fix this, added some attempts to get the service URL.
Now, the system is checking and waiting for the ceph services url
to initialization continue.

Test Plan:
PASS: Check if the message "Service url successfully retrieved"
      is displayed after "Could not get service url" in the
      ceph-manager logs.

Closes-Bug: 1997758

Signed-off-by: Gustavo Ornaghi Antunes <gustavo.ornaghiantunes@windriver.com>
Change-Id: I7d47d9c88da1afd909eb65ce510c0a562e6912b6